Sprint 3 - Parte do Gallego

+++ src/modules/questions/services/DeleteQuestionsService.ts
@@ -2,13 +2,9 @@ import { inject, injectable } from 'tsyringe';
 
 import { Questions } from '@prisma/client';
 
-// import AppError from '@shared/errors/AppError';
+import AppError from '@shared/errors/AppError';
 
 import IQuestionsRepository from '../repositories/IQuestionsRepository';
-
-interface IRequest {
-  id: string;
-}
 
 @injectable()
 export default class DeleteQuestionsService {
@@ -17,9 +13,11 @@ export default class DeleteQuestionsService {
     private questionsRepository: IQuestionsRepository,
   ) { }
 
-  public async execute({
-    id,
-  }: IRequest): Promise<Questions> {
+  public async execute(id: string): Promise<Questions> {
+    const questionExists = await this.questionsRepository.findByID(id);
+
+    if (!questionExists) throw new AppError('A question with this Id does not exist');
+
     const question = await this.questionsRepository.delete(id);
 
     return question;
